// src/types.ts
export interface MemoryPhoto {
  id: string;
  url: string;
  caption: string;
}

export interface Wish {
  id: number;
  text: string;
  emoji: string;
  color: string;
}

export interface FunnyCard {
  id: string;
  title: string;
  subtitle: string;
  emoji: string;
  desc: string;
  color: string;
}